refactor: move UseCase to core/libs
