+++ app/api/core/libs/UseCase.ts
@@ -4,8 +4,8 @@ import { EventsBus } from 'api/eventsbus';
 import { JobsDispatcher } from 'api/core/libs/queue/application/contracts/JobsDispatcher';
 import { UserSchema } from 'shared/types/userType';
 import { Tenant } from 'api/tenants/tenantContext';
-import { TransactionManager } from './TransactionManager';
-import { IdGenerator } from './IdGenerator';
+import { TransactionManager } from '../../common.v2/contracts/TransactionManager';
+import { IdGenerator } from '../../common.v2/contracts/IdGenerator';
 
 interface UseCase<Input, Output> {
   execute(input: Input, ...args: any): Promise<Output>;
