+++ app/api/paragraphExtraction/application/PXCreateParagraph.ts
@@ -1,6 +1,6 @@
 import { ObjectId } from 'mongodb';
 
-import { UseCase } from 'api/common.v2/contracts/UseCase';
+import { UseCase } from 'api/core/libs/UseCase';
 import { EntitySchema } from 'shared/types/entityType';
 import { ArrayUtils } from 'api/common.v2/utils/Array';
 import { Logger } from 'api/log.v2/contracts/Logger';
