+++ app/api/paragraphExtraction/application/PXCreateParagraphs.ts
@@ -1,6 +1,6 @@
 import { ObjectId } from 'mongodb';
 
-import { UseCase } from 'api/common.v2/contracts/UseCase';
+import { UseCase } from 'api/core/libs/UseCase';
 import { ArrayUtils } from 'api/common.v2/utils/Array';
 import entities from 'api/entities';
 import { DefaultLogger } from 'api/log.v2/infrastructure/StandardLogger';
