+++ app/api/paragraphExtraction/application/PXGetEntityParagraphs.ts
@@ -1,4 +1,4 @@
-import { UseCase } from 'api/common.v2/contracts/UseCase';
+import { UseCase } from 'api/core/libs/UseCase';
 import { SettingsDataSource } from 'api/settings.v2/contracts/SettingsDataSource';
 import {
   GetExtractedParagraphsOutput,
