+++ app/api/paragraphExtraction/application/PXGetExtractorStatuses.ts
@@ -1,6 +1,6 @@
 import { LanguageISO6391 } from 'shared/types/commonTypes';
 
-import { UseCase } from 'api/common.v2/contracts/UseCase';
+import { UseCase } from 'api/core/libs/UseCase';
 import { SettingsDataSource } from 'api/settings.v2/contracts/SettingsDataSource';
 import { FilesDataSource } from 'api/files.v2/contracts/FilesDataSource';
