+++ app/api/suggestions/useCases/updateSuggestionsAfterEntityUpdate.ts
@@ -1,4 +1,4 @@
-import { UseCase } from 'api/common.v2/contracts/UseCase';
+import { UseCase } from 'api/core/libs/UseCase';
 import { EntitySchema } from 'shared/types/entityType';
 import { IXServices } from 'api/services/informationextraction/IXServices';
 import { IXSuggestionType } from 'shared/types/suggestionType';
